% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\docType{data}
\name{Async}
\alias{Async}
\title{Simple async client}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Async
}
\description{
Simple async client
}
\examples{
\dontrun{
cc <- Async$new(
  urls = c(
    'http://localhost:9000/get',
    'http://localhost:9000/get?a=5',
    'http://localhost:9000/get?foo=bar'
  )
)
cc
(res <- cc$get())
res[[1]]
res[[1]]$url
res[[1]]$success()
res[[1]]$status_http()
res[[1]]$response_headers
res[[1]]$method
res[[1]]$content
res[[1]]$parse("UTF-8")

lapply(res, function(z) z$parse("UTF-8"))
}
}
\keyword{datasets}

