% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\docType{data}
\name{AsyncClient}
\alias{AsyncClient}
\title{Async client}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AsyncClient
}
\description{
Async client
}
\examples{
\dontrun{
cc <- AsyncClient$new(
  urls = c(
    'https://www.heroku.com/',
    'http://docs.python-tablib.org/en/latest/',
    'https://httpbin.org'
  )
)
cc
(res <- cc$get())
res[[1]]
res[[1]]$url
res[[1]]$success()
res[[1]]$status_http()
res[[1]]$response_headers
res[[1]]$method
res[[1]]$content
res[[1]]$parse("UTF-8")
}
}
\keyword{datasets}

