% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asyncvaried.R
\docType{data}
\name{AsyncVaried}
\alias{AsyncVaried}
\title{Async client for different request types}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AsyncVaried
}
\description{
Async client for different request types
}
\examples{
\dontrun{
# 1st request - GET
(req1 <- HttpClient2$new(
  url = "http://localhost:9000/get"
))
req1$get()

# 2nd request = POST
req2 <- HttpClient2$new(
  url = "http://localhost:9000/post"
)
req2$post()

# build async client
(manyreq <- AsyncVaried$new(req1, req2))
manyreq$request()
manyreq$parse()

# all in one call
req1 <- HttpClient2$new(url = "http://localhost:9000/get")$get()
req2 <- HttpClient2$new(url = "http://localhost:9000/post")$post()
out <- AsyncVaried$new(req1, req2)
out$request()
out$status()
out$status_code()
out$content()
out$times()
out$parse()

# lots of calls
reqlist <- list(
  HttpClient2$new(url = "http://localhost:9000/get")$get(),
  HttpClient2$new(url = "http://localhost:9000/post")$post(),
  HttpClient2$new(url = "http://localhost:9000/put")$put(),
  HttpClient2$new(url = "http://localhost:9000/delete")$delete(),
  HttpClient2$new(url = "http://localhost:9000/get?g=5")$get(),
  HttpClient2$new(url = "http://localhost:9000/post")$post(body = list(y = 9)),
  HttpClient2$new(url = "http://localhost:9000/get")$get(query = list(hello = "world"))
)

out <- AsyncVaried$new(.list = reqlist)
out$request()
out$status()
out$status_code()
out$content()
out$times()
out$parse()
}
}
\keyword{datasets}

