% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crul-package.r
\docType{package}
\name{crul-package}
\alias{crul}
\alias{crul-package}
\title{\strong{HTTP R client}}
\description{
\strong{HTTP R client}
}
\section{Package API}{

\itemize{
 \item \code{\link{HttpClient}} - the workhorse, and main thing to use
 \item \code{\link{HttpResponse}} - mostly for internal use, but you can use
 if you want
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}

