% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete-requests.R
\name{delete-requests}
\alias{delete-requests}
\title{HTTP DELETE requests}
\description{
HTTP DELETE requests
}
\examples{
(x <- HttpClient$new(url = "https://httpbin.org"))

## a list
(res1 <- x$delete('delete', body = list(hello = "world"), verbose = TRUE))
jsonlite::fromJSON(res1$parse("UTF-8"))

## a string
(res2 <- x$delete('delete', body = "hello world", verbose = TRUE))
jsonlite::fromJSON(res2$parse("UTF-8"))

## empty body request
x$delete('delete', verbose = TRUE)

}
