% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{handle}
\alias{handle}
\title{Make a handle}
\usage{
handle(url, ...)
}
\arguments{
\item{url}{(character) A url. required.}

\item{...}{options passed on to \code{\link[curl:new_handle]{curl::new_handle()}}}
}
\description{
Make a handle
}
\examples{
handle("https://httpbin.org")

# handles - pass in your own handle
h <- handle("https://httpbin.org")
(res <- HttpClient$new(handle = h))
out <- res$get("get")
}
