% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\docType{data}
\name{HttpClient}
\alias{HttpClient}
\title{HTTP client}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HttpClient
}
\arguments{
\item{url}{A url}

\item{opts}{curl options}

\item{handle}{A handle}
}
\description{
HTTP client
}
\details{
\strong{Methods}
  \describe{
    \item{\code{register_stub(stub)}}{
      Register a stub
    }
  }
}
\examples{
\dontrun{
x <- HttpClient$new(url = "https://httpbin.org")
x$url
(res_get1 <- x$get('get'))
res_get1$content
res_get1$response_headers
res_get1$parse()

(res_get2 <- x$get('get', query = list(hello = "world")))
res_get2$parse()
library("jsonlite")
jsonlite::fromJSON(res_get2$parse())

# post request
(res_post <- x$post('post', body = list(hello = "world")))

## empty body request
x$post('post')

# put request
(res_put <- x$put('put'))

# delete request
(res_delete <- x$delete('delete'))

# patch request
(res_patch <- x$patch('patch'))

# head request
(res_head <- x$head())
}
}
\keyword{internal}

