% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httprequest.R
\docType{data}
\name{HttpRequest}
\alias{HttpRequest}
\title{HTTP client}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HttpRequest
}
\arguments{
\item{url}{(character) A url. One of \code{url} or \code{handle} required.}

\item{opts}{(list) curl options}

\item{proxies}{an object of class \code{proxy}, as returned from the
\code{\link{proxy}} function. Supports one proxy for now}

\item{method}{(character) HTTP method: head, get, post, put, patch,
delete, options}

\item{handle}{A handle, see \code{\link{handle}}}
}
\description{
HTTP client
}
\examples{
(x <- HttpRequest$new(url = "https://httpbin.org")$get())
x$method
x$url
x$payload

(x <- HttpRequest$new(url = "http://localhost:9000/post"))
x$post(body = list(foo = "bar"))
}
\seealso{
\code{\link{post-requests}}, \code{\link{http-headers}},
\code{\link{writing-options}}
}
\keyword{datasets}

