% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\docType{data}
\name{HttpResponse}
\alias{HttpResponse}
\title{HTTP response}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HttpResponse
}
\arguments{
\item{url}{A url}

\item{opts}{curl options}

\item{handle}{A handle}
}
\description{
HTTP response
}
\details{
\strong{Methods}
  \describe{
    \item{\code{register_stub(stub)}}{
      Register a stub
    }
  }
}
\examples{
\dontrun{
x <- HttpResponse$new(method = "get", url = "https://httpbin.org")
x$url
x$method
}
}
\keyword{internal}

