% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocking.R
\name{mocking}
\alias{mocking}
\title{Mocking HTTP requests}
\description{
Mocking HTTP requests
}
\examples{
# load webmockr
library(webmockr)

# turn on mocking
webmockr::enable()
crul::mock()

# create an HTTP client
(x <- HttpClient$new(url = "http://localhost:9000"))
x$get('get')

# set headers
(res <- HttpClient$new(
  url = "https://httpbin.org",
  opts = list(
    verbose = TRUE
  ),
  headers = list(
    a = "stuff",
    b = "things"
  )
))
res$headers
# reassign header value
res$headers$a <- "that"
# define new header
res$headers$c <- "what"
# request
res$get('get')

## setting content-type via headers
(res <- HttpClient$new(
  url = "https://httpbin.org",
  opts = list(
    verbose = TRUE
  ),
  headers = list(`Content-Type` = "application/json")
))
res$get('get')
}

