% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\title{progress bars}
\description{
progress bars
}
\details{
pass \code{httr::progress()} to \code{progress} param in \link{HttpClient},
which pulls out relevant info to pass down to \pkg{curl}

if file sizes known you get progress bar; if file sizes not
known you get bytes downloaded

See the README for examples
}
