% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crul-package.r
\docType{package}
\name{crul-package}
\alias{crul}
\alias{crul-package}
\title{\strong{HTTP R client}}
\description{
\strong{HTTP R client}
}
\section{Package API}{

\itemize{
 \item \code{\link{HttpClient}} - create a connection client, set all
 your http options, make http requests
 \item \code{\link{HttpResponse}} - mostly for internal use, handles
 http responses
}
}

\section{HTTP conditions}{

We use \code{fauxpas} if you have it installed for handling HTTP
conditions but if it's not installed we use \pkg{httpcode}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}

