% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\docType{data}
\name{HttpClient}
\alias{HttpClient}
\title{HTTP client}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HttpClient
}
\arguments{
\item{url}{A url}

\item{opts}{curl options}

\item{handle}{A handle}
}
\description{
HTTP client
}
\details{
\strong{Methods}
  \describe{
    \item{\code{register_stub(stub)}}{
      Register a stub
    }
  }
}
\examples{
\dontrun{
x <- HttpClient$new(url = "https://httpbin.org")
x$url
(res <- x$get('get'))
res$content
res$response_headers
res$parse()

(res2 <- x$get('get', query = list(hello = "world")))
res2$parse()
library("jsonlite")
jsonlite::fromJSON(res2$parse())
}
}
\keyword{internal}

