% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\docType{data}
\name{HttpResponse}
\alias{HttpResponse}
\title{Base response object}
\arguments{
\item{url}{(character) A url}

\item{opts}{(list) curl options}

\item{handle}{A handle}
}
\description{
Base response object
}
\details{
\strong{Methods}
  \describe{
    \item{\code{parse()}}{
      Parse the raw response content to text
    }
  }
}
\examples{
\dontrun{
x <- HttpResponse$new(method = "get", url = "https://httpbin.org")
x$url
x$method
}
}
\keyword{datasets}

