% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocking.R
\name{mock}
\alias{mock}
\title{Mocking HTTP requests}
\usage{
mock(on = TRUE)
}
\arguments{
\item{on}{(logical) turn mocking on with \code{TRUE} or turn off with \code{FALSE}.
By default is \code{FALSE}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Mocking is now controlled by params within the various clients:
\code{\link[=HttpClient]{HttpClient()}}, \code{\link[=Async]{Async()}} and \code{\link[=AsyncVaried]{AsyncVaried()}}
}
