% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{upload}
\alias{upload}
\title{upload file}
\usage{
upload(path, type = NULL)
}
\arguments{
\item{path}{(character) a single path, file must exist}

\item{type}{(character) a file type, guessed by \link[mime:guess_type]{mime::guess_type} if
not given}
}
\description{
upload file
}
