% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{verb-DELETE}
\alias{verb-DELETE}
\title{HTTP verb info: DELETE}
\description{
The DELETE method deletes the specified resource.
}
\section{The DELETE method}{

The DELETE method requests that the origin server remove the
association between the target resource and its current
functionality.  In effect, this method is similar to the rm command
in UNIX: it expresses a deletion operation on the URI mapping of the
origin server rather than an expectation that the previously
associated information be deleted.
}

\examples{
\dontrun{
x <- HttpClient$new(url = "https://hb.opencpu.org")
x$delete(path = 'delete')

## a list
(res1 <- x$delete('delete', body = list(hello = "world"), verbose = TRUE))
jsonlite::fromJSON(res1$parse("UTF-8"))

## a string
(res2 <- x$delete('delete', body = "hello world", verbose = TRUE))
jsonlite::fromJSON(res2$parse("UTF-8"))

## empty body request
x$delete('delete', verbose = TRUE)
}

}
\references{
\url{https://datatracker.ietf.org/doc/html/rfc7231#section-4.3.5}
}
\seealso{
\link{crul-package}

Other verbs: 
\code{\link{verb-GET}},
\code{\link{verb-HEAD}},
\code{\link{verb-PATCH}},
\code{\link{verb-POST}},
\code{\link{verb-PUT}}
}
\concept{verbs}
